+++ src/components/ServerModal.vue
@@ -114,6 +114,109 @@
           </div>
         </div>
 
+        <!-- API Key Section -->
+        <div v-if="canManageKeys" class="px-6 py-4 border-b border-main-600">
+          <div class="flex items-center justify-between mb-4">
+            <h3 class="text-lg font-medium text-gray-100">API Key Management</h3>
+            <button
+              @click="toggleApiKeySection"
+              class="px-3 py-1.5 text-sm font-medium text-gray-100 bg-blue-700 hover:bg-blue-600 rounded-md transition-colors"
+            >
+              <svg class="w-4 h-4 inline mr-1" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                      :d="showApiKeySection ? 'M19 9l-7 7-7-7' : 'M9 5l7 7-7 7'"/>
+              </svg>
+              {{ showApiKeySection ? 'Hide' : 'Show' }} API Key
+            </button>
+          </div>
+
+          <!-- API Key Details (Should this be collapsible?) -->
+          <div v-if="showApiKeySection" class="space-y-4">
+            <!-- Loading State -->
+            <div v-if="loadingKeyInfo" class="flex items-center justify-center py-8">
+              <loadWheel class="w-6 h-6 mr-2"></loadWheel>
+              <span class="text-gray-400">Loading API key information...</span>
+            </div>
+
+            <!-- Key Info -->
+            <div v-else-if="currentServerKeyInfo" class="space-y-4">
+              <!-- API Key Display -->
+              <div>
+                <label class="block text-sm font-medium text-gray-300 mb-2">API Key</label>
+                <div class="relative">
+                  <div
+                    @click="copyApiKey"
+                    class="w-full px-3 py-2 bg-main-900 border border-main-600 rounded-md cursor-pointer hover:bg-main-800 transition-colors group"
+                    title="Click to copy API key"
+                  >
+                    <code class="text-sm text-gray-100 monospace blur-sm">{{ currentServerKeyInfo.key }}</code>
+                  </div>
+                  <div class="absolute inset-0 flex items-center justify-center pointer-events-none">
+                    <span class="text-xs text-gray-400 bg-main-800 px-2 py-1 rounded border border-main-600 group-hover:text-gray-300 transition-colors">
+                      <svg class="w-3 h-3 inline mr-1" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M8 16H6a2 2 0 01-2-2V6a2 2 0 012-2h8a2 2 0 012 2v2m-6 12h8a2 2 0 002-2v-8a2 2 0 00-2-2h-8a2 2 0 00-2 2v8a2 2 0 002 2z"/>
+                      </svg>
+                      Click to copy API key
+                    </span>
+                  </div>
+                </div>
+              </div>
+
+              <!-- Permissions Form -->
+              <form @submit.prevent="updateKey">
+                <div class="mb-4">
+                  <label class="block text-sm font-medium text-gray-300 mb-2">Permissions</label>
+                  <p class="text-xs text-gray-400 mb-3">Select the permissions for this server's API key.</p>
+                  <div class="space-y-2">
+                    <label class="flex items-center">
+                      <input
+                        v-model="editKeyForm.submitTimes"
+                        type="checkbox"
+                        class="rounded border-main-600 text-green-500 bg-main-900 focus:outline-none focus:ring-0"
+                        :disabled="updatingKey"
+                      />
+                      <span class="ml-2 text-sm text-gray-200">Submit Times</span>
+                    </label>
+                    <label class="flex items-center">
+                      <input
+                        v-model="editKeyForm.submitBulk"
+                        type="checkbox"
+                        class="rounded border-main-600 text-green-500 bg-main-900 focus:outline-none focus:ring-0"
+                        :disabled="updatingKey"
+                      />
+                      <span class="ml-2 text-sm text-gray-200">Submit Bulk</span>
+                    </label>
+                  </div>
+                </div>
+
+                <div class="flex justify-between items-center">
+                  <button
+                    type="button"
+                    @click="confirmDeleteKey"
+                    :disabled="deletingKey"
+                    class="px-4 py-2 text-sm font-medium text-gray-100 bg-red-700 hover:bg-red-600 rounded-md transition-colors disabled:opacity-50 cursor-pointer disabled:cursor-not-allowed"
+                  >
+                    {{ deletingKey ? 'Deleting...' : 'Delete Key' }}
+                  </button>
+
+                  <button
+                    type="submit"
+                    :disabled="updatingKey"
+                    class="px-4 py-2 text-sm font-medium text-gray-100 bg-blue-700 hover:bg-blue-600 rounded-md transition-colors disabled:opacity-50 cursor-pointer disabled:cursor-not-allowed"
+                  >
+                    {{ updatingKey ? 'Updating...' : 'Update Permissions' }}
+                  </button>
+                </div>
+              </form>
+            </div>
+
+            <!-- No Key Found -->
+            <div v-else class="text-center py-8">
+              <p class="text-gray-400 text-sm">No API key found for this server.</p>
+            </div>
+          </div>
+        </div>
+
         <!-- Servers Section -->
         <div class="px-6 py-4">
           <div class="flex items-center justify-between mb-4">
@@ -268,6 +371,43 @@
       </div>
     </div>
 
+    <!-- Delete API Key Confirmation Modal -->
+    <div v-if="showDeleteConfirm" class="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-60" @click="cancelDeleteKey">
+      <div class="bg-main-800 border border-red-400 rounded-lg shadow-lg p-6 w-[500px] max-w-[90vw] mx-4 relative" @click.stop>
+        <!-- Close X Button -->
+        <button
+          @click="cancelDeleteKey"
+          class="absolute top-4 right-4 text-gray-400 hover:text-gray-200 transition-colors cursor-pointer"
+          title="Close"
+        >
+          <svg class="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M6 18L18 6M6 6l12 12"/>
+          </svg>
+        </button>
+
+        <h3 class="text-lg font-semibold mb-4 text-red-400">⚠️ Delete API Key</h3>
+        <div class="mb-4">
+          <p class="text-sm text-gray-300 mb-2">
+            Are you sure you want to delete the API key for
+            <span class="font-semibold text-gray-100">{{ currentServerKeyInfo?.server }}</span>?
+          </p>
+          <p class="text-xs text-red-400">
+            This action cannot be undone. The server will lose access immediately and a new key will need to be created.
+          </p>
+        </div>
+        <div class="flex justify-end space-x-3">
+          <button
+            type="button"
+            @click="executeDeleteKey"
+            :disabled="deletingKey"
+            class="px-4 py-2 text-sm font-medium text-gray-100 bg-red-700 hover:bg-red-600 rounded-md transition-colors disabled:opacity-50 cursor-pointer disabled:cursor-not-allowed"
+          >
+            {{ deletingKey ? 'Deleting...' : 'Yes, Delete Key' }}
+          </button>
+        </div>
+      </div>
+    </div>
+
     <!-- Toast Notification -->
     <Toast
       :show="showToast"
@@ -281,9 +421,12 @@
 
 <script setup lang="ts">
 import { ref, computed, watch } from 'vue';
-import OffstylesApi, { type ServerDataDocument, type ServerInfo } from '@/api/offstylesApi';
+import OffstylesApi, { type ServerDataDocument, type ServerInfo, type ServerKeyInfo } from '@/api/offstylesApi';
+import { KeyPermissions, addPermission } from '@/utils/permissions';
+import { canManageApiKeys } from '@/utils/userPermissions';
 import { useAuth } from '@/stores/auth';
 import Toast from '@/components/Toast.vue';
+import loadWheel from '@/components/icons/loadWheel.vue';
 
 // Props
 interface Props {
@@ -309,6 +452,11 @@ const canEdit = computed(() => {
   if (!user.value) return false;
   // Check if user is the owner or has admin permissions
   return user.value.steam_id === props.server.user.steam_id || user.value.permissions > 0;
+});
+
+// Check if user has admin permissions for API key management
+const canManageKeys = computed(() => {
+  return user.value && canManageApiKeys(user.value.permissions);
 });
 
 // Reactive data
@@ -343,6 +491,20 @@ const showToast = ref(false);
 const toastType = ref<'success' | 'error' | 'warning' | 'info'>('info');
 const toastTitle = ref('');
 const toastMessage = ref('');
+
+// API Key management state
+const showApiKeySection = ref(false);
+const currentServerKeyInfo = ref<ServerKeyInfo | null>(null);
+const loadingKeyInfo = ref(false);
+const updatingKey = ref(false);
+const deletingKey = ref(false);
+const showDeleteConfirm = ref(false);
+
+const editKeyForm = ref({
+  server: props.server.name,
+  submitTimes: false,
+  submitBulk: false
+});
 
 // Watch for server prop changes
 watch(() => props.server, (newServer) => {
@@ -504,6 +666,112 @@ const copyToClipboard = async (text: string) => {
     document.execCommand('copy');
     document.body.removeChild(textArea);
     showToastMessage('success', 'Copied!', `${text} copied to clipboard`);
+  }
+};
+
+// API Key management functions
+const loadApiKeyInfo = async () => {
+  if (!props.server.name) return;
+
+  try {
+    loadingKeyInfo.value = true;
+    const keyInfo = await OffstylesApi.getServerKeyInfo(props.server.name);
+    currentServerKeyInfo.value = keyInfo;
+
+    // Set permissions checkboxes based on current key permissions
+    editKeyForm.value.submitTimes = (keyInfo.permissions & KeyPermissions.SUBMIT_TIMES) !== 0;
+    editKeyForm.value.submitBulk = (keyInfo.permissions & KeyPermissions.SUBMIT_BULK) !== 0;
+  } catch (err) {
+    console.error('Failed to load API key info:', err);
+    currentServerKeyInfo.value = null;
+  } finally {
+    loadingKeyInfo.value = false;
+  }
+};
+
+const toggleApiKeySection = async () => {
+  showApiKeySection.value = !showApiKeySection.value;
+
+  if (showApiKeySection.value && !currentServerKeyInfo.value && !loadingKeyInfo.value) {
+    await loadApiKeyInfo();
+  }
+};
+
+const updateKey = async () => {
+  try {
+    updatingKey.value = true;
+
+    let permissions = 0;
+    if (editKeyForm.value.submitTimes) {
+      permissions = addPermission(permissions, KeyPermissions.SUBMIT_TIMES);
+    }
+    if (editKeyForm.value.submitBulk) {
+      permissions = addPermission(permissions, KeyPermissions.SUBMIT_BULK);
+    }
+
+    await OffstylesApi.updateApiKey(
+      editKeyForm.value.server,
+      permissions
+    );
+
+    // Reload key info to get updated data
+    await loadApiKeyInfo();
+
+    showToastMessage('success', 'Success', 'API key permissions updated successfully');
+  } catch (err) {
+    showToastMessage('error', 'Error', 'Failed to update API key permissions');
+    console.error('Failed to update API key permissions:', err);
+  } finally {
+    updatingKey.value = false;
+  }
+};
+
+const confirmDeleteKey = () => {
+  showDeleteConfirm.value = true;
+};
+
+const cancelDeleteKey = () => {
+  showDeleteConfirm.value = false;
+};
+
+const executeDeleteKey = async () => {
+  if (!currentServerKeyInfo.value) return;
+
+  try {
+    deletingKey.value = true;
+
+    await OffstylesApi.deleteApiKey(currentServerKeyInfo.value.server);
+
+    // Clear the key info and close the section
+    currentServerKeyInfo.value = null;
+    showApiKeySection.value = false;
+    showDeleteConfirm.value = false;
+
+    showToastMessage('success', 'Deleted!', `API key for ${props.server.name} has been deleted`);
+  } catch (err) {
+    showToastMessage('error', 'Error', 'Failed to delete API key');
+    console.error('Failed to delete API key:', err);
+  } finally {
+    deletingKey.value = false;
+  }
+};
+
+const copyApiKey = async () => {
+  if (currentServerKeyInfo.value?.key) {
+    try {
+      await navigator.clipboard.writeText(currentServerKeyInfo.value.key);
+      showToastMessage('success', 'Copied!', `API key for ${props.server.name} copied to clipboard`);
+    } catch (err) {
+      console.error('Failed to copy API key to clipboard:', err);
+      // Fallback for older browsers
+      const textArea = document.createElement('textarea');
+      textArea.value = currentServerKeyInfo.value.key;
+      document.body.appendChild(textArea);
+      textArea.select();
+      document.execCommand('copy');
+      document.body.removeChild(textArea);
+      showToastMessage('success', 'Copied!', `API key for ${props.server.name} copied to clipboard`);
+    }
   }
 };
